+++ middlewares/content-security-policy/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Removed
 
 - Drop support for Node 10 and 11. Node 12+ is now required
+- `useDefaults` option now defaults to `true`
 
 ## 3.4.0 - 2021-05-02